refactor(mcp): split server.py (1174 LOC) into tools/ subpackage (#193)
`mcp/server.py` was a flat ~1200-line module mixing the FastMCP singleton,
14 tool functions across 5 unrelated domains, and shared helpers. Splits
along the domain seams the original file's section comments already
described, so the layout is mechanical rather than design-by-fiat:

  src/srunx/mcp/
  ├── app.py             FastMCP singleton + missing-package guard
  ├── helpers.py         ok/err/validate_*/job_to_dict/get_ssh_client/reject_python_prefix
  ├── server.py          27-line entry point: imports tools, runs mcp.run()
  └── tools/
      ├── jobs.py        submit_job/list_jobs/get_job_status/cancel_job/get_job_logs
      ├── resources.py   get_resources
      ├── workflows.py   create/validate/run/list/get_workflow + 2 internal guards
      ├── sync.py        sync_files
      └── config.py      get_config/list_ssh_profiles

The singleton lives in `app.py` (not `server.py`) so tool modules can do
`from srunx.mcp.app import mcp` without depending on the entry point;
otherwise tool registration would create a circular import.

Helper renames (drop `_` prefix on cross-module helpers):
  _ok / _err                          → ok / err
  _validate_job_id / _validate_partition → validate_job_id / validate_partition
  _job_to_dict                        → job_to_dict
  _get_ssh_client                     → get_ssh_client
  _reject_python_prefix_mcp           → reject_python_prefix

Workflow-internal guards (`_enforce_shell_script_roots`, `_resolve_mount_context`)
stay private to `tools/workflows.py` since they reference workflow-layer
types and have no other consumer. The `_mcp` suffix on
`_enforce_shell_script_roots_mcp` was redundant inside `mcp/tools/` —
dropped.

Tests mirror the new layout (CLAUDE.md test layout rules #1):

  tests/mcp/
  ├── test_helpers.py
  └── tools/
      ├── test_jobs.py / test_resources.py / test_workflows.py / test_config.py
      └── (sync.py has no tests yet — same as before split)

`tests/mcp/test_server.py` (962 LOC) and `tests/mcp/test_sweep.py` (361
LOC) were broken up; git detected the workflow-tests rename so sweep test
history follows to `tools/test_workflows.py`. `@patch` paths updated to
target the new lookup sites (`srunx.mcp.tools.jobs.get_ssh_client` etc.)
per the standard "patch where it's used" rule.

Side cleanups (eliminating quickfix-ratchet deferrals):
  - dropped `# type: ignore` on `import yaml` — `types-pyyaml` is in dev deps
  - replaced `except Exception: # noqa: BLE001` with `except (ValueError,
    TypeError)` (Pydantic v2 ValidationError extends ValueError; covers
    int() conversion + None paths)
  - rejected the `# noqa: F401` pattern for side-effect-only imports —
    bind imports to `_REGISTERED_TOOL_MODULES` tuple instead so ruff sees
    them as used

Verified: 14 tools registered (same as pre-split), 87 mcp tests pass,
2042 total tests pass, ruff + mypy clean.

Refs #193 (3 of 4 module splits remaining: cli/workflow.py,
ssh/core/client.py, slurm/ssh.py).
